enhance import
If card exists in srf (UNIQUE constraint violation on creating card),
add the card to the map. This occurs when importing a collection
multiple times.

Only create revlog records if there is a corresponding card. This
discards history for cards that are no longer in the deck. Presumably
the cards existed at some point in the past, but they no longer exist.