Reduce server state
The server was keeping track of the card being viewed.

With this change, the server has no record of a card being viewed. When
a card is reviewed, the state is in the client / browser, and the
response must include all details of the review, including the time
spent viewing the card.

This admits multiple concurrent browser sessions.

Use of the browser back button may still be problematic because that
will return to the original start time.