Change default previewWindow to 0
After trying it for a while, I don't think it is a good idea. Viewing
cards without delay doesn't aid learning and one can simply not study
for a while. There is little value in exercising short term memory and
many pointless repititions to set a longer interval. Just stop studying,
if you want a break.