Add a button to get a new card.
Normally, a new card is only presented if various constraints are
satisfied. This is to automatically regulate workload. But the user
might want to see new cards despite these limits being exceeded (e.g.
despite average study time exceeding the configured limit or despite the
number of new cards viewed today exceeding the configured limit). To this
end, rather than adjusting the limits, the user can simply click the New
Card button to get a new card and view as many new cards as desired.

In passing, reorganized the buttons at the bottom of the home page and
removed the Refresh button as the link at the top of the page or browser
refresh can be used to refresh the home page.