Change estimates of future study time
The old estimates were significantly inaccurate. Change the heuristic
for estimating future study time to be simpler and more accurate. Remove
the estimate of total study time for the day from the stats page as
calculating this accurately is particularly complicated. The inaccurate
estimate doesn't help and it is too difficult to produce an accurate
estimate. Change estimate of time per card to time per review.