Change interval calculations
Next interval was calculated on the basis of actual last interval but
for short intervals the actual last interval is heavily influenced by
breaks in study (e.g. one study session per day or morning and evening
sessions). These breaks create clusters of intervals. It's not
necessarily wrong or ineffective. Breaks in study are real and after a
break the actual interval is what it is even if it wasn't planned. None
the less, with this change, try ignoring the actual interval at short
intervals and use the scheduled interval instead. Short is deemed one
week so that a 24 hour break is about 10%.