Remote the templateset table.
It was reduced to just recording the name. Instead of a separate table,
linked by ID, the fieldset and template tables have been changed to
store the templateset name in field templateset.

The templateset still exists: it is the set of templates with the same
templateset value.

Fieldsets are still linked to templatesets, by way of their templateset
value, holding the name of the templateset.

This change makes it impossible to have multiple templatesets with the
same name, which is a good thing.