Change intervalGood
If interval exceeds the learning threshold, use the average of scheduled
and actual intervals to calculate the new interval. Not quite as
aggressive as using the actual interval, in the case of long delay, but
not as tedious as always using the scheduled interval. Though,
practically, it makes little difference as the learning threshold is 7
days and study gaps of more than a day or two are quite rare. So, the
interval used will almost always be quite close to the scheduled
interval.