Improve Study Time per Hour chart
The estimation of time per card for cards due in the next 24 hours was
overly complicated and probably not accurate. The new implementation
considers the past 100 hours with study and averages the time per unique
card in each hour. This should account for cards that are studied
multiple times in an hour (very new cards).