Change calculation of mature card count for dailystats
Preious calculation was sumulative sum of (matured - lapsed) but cards
that are deleted or reset are not counted as lapsed. As a result, the
count of matured cards tended to be larger than the number of cards
that currently have a mature threshold. After a long study history with
many resets and edits of fieldsets and templatesets, the discrepancy can
become significant. With this change, only cards that currently exist
and have an interval greater than zero are considered, when calculating
mature card count.

It is not wrong to consider the number of cards that had a mature
threshold when they were last reviewed regardless of subsequent resets
and deletions (the old algorithm) but after resets and deletions of
cards, it is surprising when the count of mature cards is greater than
the number of cards currently existing with interval greater than the
mature threshold and even greater than the count of currently existing
cards with interval greater than zero.